requirements.txt

asgiref==3.4.1
boto3==1.20.7
botocore==1.23.7
crispy==0.6.3
csp==0.1a10
cycler==0.11.0
Django==3.2.9
django-admin-rangefilter == 0.8.2
django-crispy-forms==1.13.0
django-daterange-filter==1.3.0
django-online-users==0.3
fabio==0.12.0
fonttools== 4.28.1
future==0.18.2
h5py==3.6.0
hashring==1.5.1
jmespath==0.10.0
kiwisolver==1.3.2
matplotlib==3.5.0
memcache==0.10.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pip==21.2.4
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5== 5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-memcached==1.59
pytz==2021.3
s3transfer==0.5.0
setuptools==58.2.0
setuptools-scm==6.3.2
silx== 0.15.2
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
urllib3==1.26.7
validate-email==1.3
wheel==0.37.0
